requirements.txt
boto3
supabase
python-dotenv
openai
tiktoken
langchain
langchain-community
langchain-openai
